move watermark test to helper suite
